+++ paleomix/nodes/samtools.py
@@ -40,6 +40,10 @@ SAMTOOLS_VERSION = versions.Requirement(call=("samtools",),
                                         search=_VERSION_REGEX,
                                         checks=_COMMON_CHECK)
 
+SAMTOOLS_VERSION_1x = versions.Requirement(call=("samtools",),
+                                           search=_VERSION_REGEX,
+                                           checks=versions.GE(1, 0, 0))
+
 SAMTOOLS_VERSION_0119 = versions.Requirement(call=("samtools",),
                                              search=_VERSION_REGEX,
                                              checks=versions.EQ(0, 1, 19))
@@ -133,26 +137,37 @@ class FastaIndexNode(CommandNode):
 class BAMIndexNode(CommandNode):
     """Indexed a BAM file using 'samtools index'."""
 
-    def __init__(self, infile, dependencies=()):
+    def __init__(self, infile, index_format='.bai', dependencies=()):
         basename = os.path.basename(infile)
+
+        if index_format == '.bai':
+            samtools_version = SAMTOOLS_VERSION
+            samtools_call = ["samtools", "index", "%(TEMP_IN_BAM)s"]
+        elif index_format == '.csi':
+            samtools_version = SAMTOOLS_VERSION_1x
+            samtools_call = ["samtools", "index", "-c", "%(TEMP_IN_BAM)s"]
+        else:
+            raise ValueError("Unknown format type %r; expected .bai or .csi"
+                             % (index_format,))
 
         cmd_link = AtomicCmd(["ln", "-s", "%(IN_BAM)s", "%(TEMP_OUT_BAM)s"],
                              IN_BAM=infile,
                              TEMP_OUT_BAM=basename,
                              set_cwd=True)
 
-        cmd_index = AtomicCmd(["samtools", "index", "%(TEMP_IN_BAM)s"],
+        cmd_index = AtomicCmd(samtools_call,
                               TEMP_IN_BAM=basename,
-                              CHECK_SAM=SAMTOOLS_VERSION)
+                              CHECK_SAM=samtools_version)
 
         cmd_rename = AtomicCmd(["mv", "%(TEMP_IN_BAM)s", "%(OUT_BAM)s"],
-                               TEMP_IN_BAM=basename + ".bai",
-                               OUT_BAM=swap_ext(infile, ".bai"))
+                               TEMP_IN_BAM=basename + index_format,
+                               OUT_BAM=swap_ext(infile, index_format))
 
         commands = SequentialCmds((cmd_link, cmd_index, cmd_rename))
 
         CommandNode.__init__(self,
-                             description="<BAMIndex: '%s'>" % (infile,),
+                             description="<BAMIndex (%s): '%s'>"
+                             % (index_format[1:].upper(), infile),
                              command=commands,
                              dependencies=dependencies)