+++ paleomix/tools/bam_pipeline/parts/prefix.py
@@ -9,8 +9,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
 #
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -29,31 +29,52 @@ from paleomix.tools.bam_pipeline.nodes import \
 from paleomix.nodes.validation import \
     DetectInputDuplicationNode
 
+import paleomix.ui as ui
 import paleomix.nodes.gatk as gatk
 
 
-class Prefix:
+class Prefix(object):
     def __init__(self, config, prefix, samples, features, target):
-        self.name      = prefix["Name"]
-        self.label     = prefix.get("Label") or self.name
-        self.reference = prefix["Reference"]
-        self.roi       = prefix.get("RegionsOfInterest", {})
+        self.name = prefix["Name"]
+        self.label = prefix.get("Label") or self.name
+        self.roi = prefix.get("RegionsOfInterest", {})
 
         self.samples = safe_coerce_to_tuple(samples)
-        self.folder  = config.destination
-        self.target  = target
+        self.folder = config.destination
+        self.target = target
 
         files_and_nodes = {}
         for sample in self.samples:
             files_and_nodes.update(sample.bams.iteritems())
 
-        self.datadup_check = self._build_dataduplication_node(prefix, files_and_nodes)
+        self.datadup_check = self._build_dataduplication_node(
+            prefix, files_and_nodes)
+
+        build_raw_bam = features["RawBAM"]
+        build_realigned_bam = features["RealignedBAM"]
+        if build_realigned_bam and prefix['IndexFormat'] == '.csi':
+            if prefix['Path'] not in _CSI_WARNINGS:
+                ui.print_err("\nWARNING: Realigned BAMs enabled for reference "
+                             "genome %r, but the file contains sequences too "
+                             "large for GATK, which does not support .csi "
+                             "index files. Raw BAMs will be built instead of "
+                             "realigned BAMs, for this reference sequence."
+                             % (prefix['Path']))
+
+                # TODO: Add reference to FAQ when written.
+
+            _CSI_WARNINGS.add(prefix['Path'])
+            build_realigned_bam = False
+            build_raw_bam = True
 
         self.bams = {}
-        if features["RawBAM"]:
-            self.bams.update(self._build_raw_bam(config, prefix, files_and_nodes))
-        if features["RealignedBAM"]:
-            self.bams.update(self._build_realigned_bam(config, prefix, files_and_nodes))
+        if build_raw_bam:
+            self.bams.update(self._build_raw_bam(
+                config, prefix, files_and_nodes))
+
+        if build_realigned_bam:
+            self.bams.update(self._build_realigned_bam(
+                config, prefix, files_and_nodes))
 
         if not self.bams:
             for sample in self.samples:
@@ -65,21 +86,29 @@ class Prefix:
         self.nodes = tuple(nodes)
 
     def _build_raw_bam(self, config, prefix, files_and_bams):
-        output_filename = os.path.join(self.folder, "%s.%s.bam" % (self.target, prefix["Name"]))
-        validated_filename = os.path.join(self.folder, self.target, prefix["Name"] + ".validated")
+        output_filename = os.path.join(
+            self.folder, "%s.%s.bam" % (self.target, prefix["Name"]))
+        validated_filename = os.path.join(
+            self.folder, self.target, prefix["Name"] + ".validated")
 
-        node = MergeSamFilesNode(config       = config,
-                                 input_bams   = files_and_bams.keys(),
-                                 output_bam   = output_filename,
-                                 dependencies = self.datadup_check)
-        validated_node = index_and_validate_bam(config, prefix, node, validated_filename)
+        node = MergeSamFilesNode(config=config,
+                                 input_bams=files_and_bams.keys(),
+                                 output_bam=output_filename,
+                                 dependencies=self.datadup_check)
+        validated_node = index_and_validate_bam(config=config,
+                                                prefix=prefix,
+                                                node=node,
+                                                log_file=validated_filename)
 
-        return {output_filename : validated_node}
+        return {output_filename: validated_node}
 
     def _build_realigned_bam(self, config, prefix, bams):
-        output_filename = os.path.join(self.folder, "%s.%s.realigned.bam" % (self.target, prefix["Name"]))
-        intervals_filename = os.path.join(self.folder, self.target, prefix["Name"] + ".intervals")
-        validated_filename = os.path.join(self.folder, self.target, prefix["Name"] + ".realigned.validated")
+        output_filename = os.path.join(
+            self.folder, "%s.%s.realigned.bam" % (self.target, prefix["Name"]))
+        intervals_filename = os.path.join(
+            self.folder, self.target, prefix["Name"] + ".intervals")
+        validated_filename = os.path.join(
+            self.folder, self.target, prefix["Name"] + ".realigned.validated")
 
         trainer = gatk.GATKIndelTrainerNode(config=config,
                                             reference=prefix["Reference"],
@@ -95,12 +124,23 @@ class Prefix:
                                               outfile=output_filename,
                                               dependencies=trainer)
 
-        validated_node = index_and_validate_bam(config, prefix, aligner, validated_filename)
+        validated_node = index_and_validate_bam(config=config,
+                                                prefix=prefix,
+                                                node=aligner,
+                                                log_file=validated_filename)
 
         return {output_filename: validated_node}
 
     def _build_dataduplication_node(self, prefix, files_and_nodes):
-        destination = os.path.join(self.folder, self.target, prefix["Name"] + ".duplications_checked")
-        return DetectInputDuplicationNode(input_files = files_and_nodes.keys(),
-                                          output_file = destination,
-                                          dependencies = files_and_nodes.values())
+        filename = prefix["Name"] + ".duplications_checked"
+        destination = os.path.join(self.folder, self.target, filename)
+        dependencies = files_and_nodes.values()
+
+        return DetectInputDuplicationNode(input_files=files_and_nodes.keys(),
+                                          output_file=destination,
+                                          dependencies=dependencies)
+
+
+# Contains the paths of sequences for which warnings about GATK has been given,
+# if the 'RealignedBAM' feature was enabled for files that require a CSI index.
+_CSI_WARNINGS = set()